requirements.txt
streamlit
reportlab
matplotlib
Pillow